Update snapshot tests
