+++ src/__tests__/snapshots/button.test.js
@@ -1,0 +1,7 @@
+import renderer from "react-test-renderer";
+import Button from "../../components/Button";
+
+it("renders Button correctly", () => {
+  const tree = renderer.create(<Button />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
