+++ src/__tests__/snapshots/display.test.js
@@ -1,0 +1,7 @@
+import renderer from "react-test-renderer";
+import Display from "../../components/Display";
+
+it("renders Display correctly", () => {
+  const tree = renderer.create(<Display />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
