+++ src/__tests__/snapshots/quote.test.js
@@ -1,0 +1,7 @@
+import renderer from "react-test-renderer";
+import Quote from "../../components/Quote";
+
+it("renders Quote correctly", () => {
+  const tree = renderer.create(<Quote />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
