+++ src/components/Navbar.js
@@ -1,19 +1,19 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React from "react";
+import { Link } from "react-router-dom";
 
 const Navbar = () => (
   <div className="navbar">
     <h3>Math Magicians</h3>
     <div className="navbar-links">
-      <Link to="/" className="navbar-link">
+      <a href="/" className="navbar-link">
         Home
-      </Link>
-      <Link to="/calculator" className="navbar-link">
+      </a>
+      <a href="/calculator" className="navbar-link">
         Calculator
-      </Link>
-      <Link to="/quote" className="navbar-link">
+      </a>
+      <a href="/quote" className="navbar-link">
         Quote
-      </Link>
+      </a>
     </div>
   </div>
 );
